Change RPC functions to mapmodule, rpc and vectorlayerplugin. Also remove created util class because util function call not work in rpc (come "type error" similar error).
